chore: not to import tests as a package
We don't want tests to be imported as a package, so no __init__.py
provided, but some test file needs to import class/functions from the
test fixture file, conftest.py.
Under this circumstance, import the conftest directly as a module.
It's good since they're all under the test/ directory.
Also "mypy" doesn't buy it when you import tests as a package.